refactor: move IconCustomizer component
